refactor: removed old xblock-content
