+++ src/course-unit/course-xblock/CourseXblock.scss
@@ -1,4 +1,4 @@
-@import "xblock-iframe/XBlockIframe";
+@import "xblock-content/XBlockIframe";
 
 .course-unit {
   .course-unit__xblocks {
