--- src/course-unit/course-xblock/hooks/index.js
@@ -1,3 +1,0 @@
-export { default as useIFrameBehavior } from './useIFrameBehavior';
-export { default as useLoadBearingHook } from './useLoadBearingHook';
-export { default as useEventListener } from './useEventListener';
