--- src/course-unit/course-xblock/urls.js
@@ -1,18 +1,0 @@
-import { getConfig } from '@edx/frontend-platform';
-
-import { objectToQueryString } from '../../utils';
-
-export const iframeParams = {
-  show_title: 0,
-  show_bookmark: 0,
-  recheck_access: 1,
-  view: 'student_view',
-  is_authoring_mfe: 1,
-};
-
-export const getIFrameUrl = ({ blockId }) => {
-  const xblockUrl = `${getConfig().LMS_BASE_URL}/xblock/${blockId}`;
-  return `${xblockUrl}?${objectToQueryString(iframeParams)}`;
-};
-
-export default { getIFrameUrl };
