+++ src/course-unit/course-xblock/xblock-content/index.js
@@ -1,2 +1,2 @@
 /* eslint-disable-next-line import/prefer-default-export */
-export { default as XBlockIframe } from './XBlockIframe';
+export { default as XBlockContent } from './XBlockContent';
