+++ webpack.prod.config.js
@@ -21,7 +21,7 @@ const config = createConfig('webpack-prod', {
 config.plugins.push(
   new CopyPlugin({
     patterns: [{
-      context: path.resolve(__dirname, 'src/course-unit/course-xblock/xblock-iframe/iframe-wrapper'),
+      context: path.resolve(__dirname, 'src/course-unit/course-xblock/xblock-content/iframe-wrapper'),
       from: 'xblock-bootstrap.html',
     }],
   }),
